Updated lock folder handling
