+++ Helper/Config/RuntimeConfig.php
@@ -520,79 +520,4 @@ class RuntimeConfig
     {
         $this->allowMultipleExecution = $allowMultipleExecution;
     }
-
-    /**
-     * Used to override the default folder where your lock-files are stored.
-     *
-     * Providing a value of NULL with result in the Symfony default of creating the lock file in the temporary directory of the system
-     * If an absolute path is provided, the directory must already exist and be accessible to the PHP process.
-     * In POSIX environments, paths can be provided that start with ~/. This will be expanded using the $HOME environment
-     * variable and the full path subjected to he same constraints as absolute paths.
-     * or ~/ we assume you have a static location for it.
-     * All other values will be considered to be relative to the Symfony Kernel root directory of your application.
-     *
-     * @param string $lockFileFolder
-     *
-     * @return $this
-     * @throws BaseCommandException
-     */
-    public function setLockFileFolder($lockFileFolder)
-    {
-        if ($this->getExecutionPhase() > self::PHASE_INITIALISE) {
-            throw new BaseCommandException('Cannot change the location of the lock file. Lock handler is already initialised');
-        }
-
-        $this->lockFileFolder = $lockFileFolder;
-        return $this;
-    }
-
-    /**
-     * Gets the folder where the lock files will be stored, performing any necessary expansions and validations in the process
-     *
-     * @return string
-     * @throws BaseCommandException
-     */
-    public function getLockFileFolder()
-    {
-        if($this->lockFileFolder == 'lock_file_folder_undefined') {
-            return null;  // TODO add warning when attempting to access value before it is set
-        }
-
-        $fs = new Filesystem();
-        if(is_null($this->lockFileFolder)){
-             return null; // Null will result in the temporary directory of the system being used by the handler
-
-        } elseif(substr($this->lockFileFolder, 0, 2) === '~/'){ // Contains tilde
-            // Expand tilde to user's home directory
-            $homeDir = getenv('HOME');
-            if($homeDir !== false){
-                $lockDirectory = $homeDir.DIRECTORY_SEPARATOR.substr($this->lockFileFolder, 2);
-                $realPath = realpath($lockDirectory);
-                if($realPath === false){
-                    throw new BaseCommandException('Lock file folders outside of the project directory will not be created '.
-                        'automatically. The provided directory does not exist or is not accessible: '.$lockDirectory);
-                }
-                return $realPath;
-            } else {
-                throw new BaseCommandException('Could not resolve tilde (~) to the user\'s home directory for the lock '.
-                    ' file folder. Please check the $HOME environment variable of the user executing the command or consider '.
-                    'using an absolute path'
-                );
-
-            }
-
-        } elseif($fs->isAbsolutePath($this->lockFileFolder)){ // Is absolute path (Cross-platform check that works on windows)
-            $realPath = realpath($this->lockFileFolder);
-            if( $realPath === false){
-                throw new BaseCommandException('Lock file folders outside of the project directory will not be created '.
-                    'automatically. The provided directory does not exist or is not accessible: '.$this->lockFileFolder);
-            }
-            return $realPath;
-
-        } else { // Relative path
-            // Prepend Kernel Root directory
-            return $this->getContainer()->get('kernel')->getRootDir() . DIRECTORY_SEPARATOR . $this->lockFileFolder;
-        }
-    }
-
 }
