+++ Helper/Locking/LockingEnhancement.php
@@ -4,11 +4,11 @@ namespace Afrihost\BaseCommandBundle\Helper\Locking;
 
 use Afrihost\BaseCommandBundle\Exceptions\BaseCommandException;
 use Afrihost\BaseCommandBundle\Exceptions\LockAcquireException;
-use Afrihost\BaseCommandBundle\Helper\AbstractEnhancement;
 use Afrihost\BaseCommandBundle\Helper\Config\RuntimeConfig;
 use Afrihost\BaseCommandBundle\Helper\EnhancementInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Filesystem\LockHandler;
 
 class LockingEnhancement implements EnhancementInterface
@@ -24,18 +24,23 @@ class LockingEnhancement implements EnhancementInterface
     private $locking;
 
     /**
-     * @var
+     * @var string
      */
     private $lockFileFolder;
 
     /**
-     * LockingHandler constructor.
-     *
-     * @param string $lockFileFolder
+     * @var string
      */
-    public function __construct($lockFileFolder)
+    private $rootDir;
+
+    /**
+     * @param string $lockFileFolder
+     * @param string $rootDir
+     */
+    public function __construct($lockFileFolder, $rootDir)
     {
         $this->lockFileFolder = $lockFileFolder;
+        $this->rootDir = $rootDir;
     }
 
     /**
@@ -47,7 +52,7 @@ class LockingEnhancement implements EnhancementInterface
     public function initialize(InputInterface $input, OutputInterface $output)
     {
         if ($this->isLocking()) {
-            $this->lockHandler = new LockHandler($input->getFirstArgument(), $this->lockFileFolder);
+            $this->lockHandler = new LockHandler($input->getFirstArgument(), $this->getLockFileFolder());
 
             if (!$this->lockHandler->lock()) {
                 throw new LockAcquireException('Sorry, can\'t get the lock. Bailing out!');
@@ -126,5 +131,62 @@ class LockingEnhancement implements EnhancementInterface
         $this->locking = $value;
 
         return $this;
+    }
+
+    /**
+     * Gets the folder where the lock files will be stored, performing any necessary expansions and validations in the process
+     *
+     * @return string
+     * @throws BaseCommandException
+     */
+    private function getLockFileFolder()
+    {
+        if (null === $this->lockFileFolder) {
+            return null;  // TODO add warning when attempting to access value before it is set
+        }
+
+        $fs = new Filesystem();
+        if(is_null($this->lockFileFolder)){
+            return null; // Null will result in the temporary directory of the system being used by the handler
+
+        } elseif(substr($this->lockFileFolder, 0, 2) === '~/'){ // Contains tilde
+            // Expand tilde to user's home directory
+            $homeDir = getenv('HOME');
+            if($homeDir !== false){
+                $lockDirectory = $homeDir.DIRECTORY_SEPARATOR.substr($this->lockFileFolder, 2);
+                $realPath = realpath($lockDirectory);
+                if($realPath === false){
+                    throw new BaseCommandException('Lock file folders outside of the project directory will not be created '.
+                        'automatically. The provided directory does not exist or is not accessible: '.$lockDirectory);
+                }
+                return $realPath;
+            } else {
+                throw new BaseCommandException('Could not resolve tilde (~) to the user\'s home directory for the lock '.
+                    ' file folder. Please check the $HOME environment variable of the user executing the command or consider '.
+                    'using an absolute path'
+                );
+
+            }
+
+        } elseif($fs->isAbsolutePath($this->lockFileFolder)){ // Is absolute path (Cross-platform check that works on windows)
+            $realPath = realpath($this->lockFileFolder);
+            if( $realPath === false){
+                throw new BaseCommandException('Lock file folders outside of the project directory will not be created '.
+                    'automatically. The provided directory does not exist or is not accessible: '.$this->lockFileFolder);
+            }
+            return $realPath;
+
+        } else { // Relative path
+            // Prepend Kernel Root directory
+            return $this->rootDir . DIRECTORY_SEPARATOR . $this->lockFileFolder;
+        }
+    }
+
+    /**
+     * @param string $locking
+     */
+    public function setLockFileFolder($locking)
+    {
+        $this->lockFileFolder = $locking;
     }
 }
